pyproject.toml
[project]
name = "my-operator"
version = "0.1.0"
description = "Local AI operator with plan-approve-execute workflow"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai>=1.50.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "websockets>=13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
